Update submenu refresh logic and add refresh property in test-menu.json

+++ UniversalNavMenuESP32S3.ino
@@ -140,7 +140,7 @@ void loop()
       inSubMenu = false;
     }
   }
-  if (inSubMenu)
+  if (inSubMenu && (menuItems["menu"][currentScreenId]["subMenu"][subMenuIndex]["refresh"] == "true"))
   {
     spriteCurrent.deleteSprite();
     spriteCurrent.createSprite(screenW, screenH);
